#10922 refactor code
